# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Raze Posting API
  version: 0.1.0
  summary: Internal API for multi-account Meta campaign publishing and tracking
  description: |
    API-only internal service for official Meta OAuth, advertising inventory
    discovery, asynchronous multi-account publishing, persisted Insights, and
    spend-checkpoint guards fed by Facebook Insights and Keitaro tracker
    statistics (registrations, deposits, revenue).

    Except for health/readiness, the served OpenAPI document, and the Meta OAuth
    callback, callers authenticate with the administrator bearer token. Meta
    user access tokens are obtained only through OAuth and are never returned.

    Instant Forms/lead retrieval, messaging destinations, and catalogs are
    outside this version's scope.
  contact:
    email: mike@watchers-factory.com
  license:
    name: Proprietary and confidential
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.terahash.win
    description: Production
  - url: http://127.0.0.1:8080
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /openapi.yaml:
    get:
      tags: [System]
      operationId: getOpenAPI
      summary: Download this API contract
      security: []
      responses:
        "200":
          description: OpenAPI 3.1 YAML.
          content:
            application/yaml:
              schema:
                type: string
            text/yaml:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
...
  /oauth/facebook/callback:
    get:
      tags: [OAuth]
      operationId: handleFacebookOAuthCallback
      summary: Consume Meta's OAuth redirect
      description: |
        Public browser callback registered in the Meta app. Replay protection
        comes from the short-lived, one-time `state` generated by
        `POST /v1/oauth/sessions`. The response never contains a Meta token.
      security: []
      parameters:
        - name: code
          in: query
          description: Authorization code returned after a successful login.
          schema:
            type: string
        - name: state
          in: query
          description: One-time opaque OAuth state.
          required: true
          schema:
            type: string
            minLength: 1
        - name: error
          in: query
          schema:
            type: string
        - name: error_reason
          in: query
          schema:
            type: string
        - name: error_description
          in: query
          schema:
            type: string
      responses:
        "200":
          description: The connection was recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthCallbackResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/oauth/sessions:
    post:
      tags: [OAuth]
      operationId: createOAuthSession
      summary: Start an official Meta OAuth session
      description: |
        Creates a 30-minute, single-use state and returns the Meta authorization
        URL to open in the intended user's browser.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOAuthSessionRequest"
      responses:
        "201":
          description: OAuth session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOAuthSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/connections:
    get:
      tags: [Connections]
      operationId: listConnections
      summary: List connected Meta users
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ConnectionStatus"
        - name: search
          in: query
          description: Case-insensitive match against name, email, or Meta user ID.
          schema:
            type: string
      responses:
        "200":
          description: Paginated connections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/connections/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Connections]
      operationId: getConnection
      summary: Get one Meta connection
      responses:
        "200":
          description: Connection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Connections]
      operationId: revokeConnection
      summary: Revoke a Meta connection and its consent grant
      description: |
        Revokes the current person's Meta grant before marking the local
        connection revoked. Reconnect through OAuth to establish a fresh
        consent grant with the currently configured permissions.
      responses:
        "204":
          description: Connection grant revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/connections/{id}/sync:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Connections]
      operationId: syncConnection
      summary: Queue inventory synchronization
      description: |
        Persists a synchronization job. A successful response means the job was
        accepted, not that Meta inventory is already current.
      responses:
        "202":
          description: Synchronization queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/ad-accounts:
    get:
      tags: [Inventory]
      operationId: listAdAccounts
      summary: List discovered ad accounts
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: connection_id
          in: query
          schema:
            type: string
            format: uuid
        - name: business_id
          in: query
          schema:
            type: string
            format: uuid
        - name: account_status
          in: query
          description: Numeric Meta account status.
          schema:
            type: integer
        - name: currency
          in: query
          description: ISO-style account currency code, for example `USD`.
          schema:
            type: string
            minLength: 3
            maxLength: 3
        - name: active_only
          in: query
          description: Return only accounts still accessible through the connection. Defaults to true.
          schema:
            type: boolean
            default: true
        - name: search
          in: query
          description: Case-insensitive name or account-ID match.
          schema:
            type: string
      responses:
        "200":
          description: Paginated ad accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccountPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/ad-accounts/{id}/campaign-audit:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Inventory]
      operationId: auditAdAccountCampaigns
      summary: Read live campaign structures from Meta
      description: |
        Fetches campaigns, ad sets, ads, and creative details directly from
        Meta for buyer-side inspection. Defaults to objects with effective
        status `ACTIVE`.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: effective_status
          in: query
          description: Comma-separated Meta effective statuses. Defaults to `ACTIVE`.
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              examples: [ACTIVE, PAUSED]
      responses:
        "200":
          description: Live Meta campaign structures for the selected ad account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccountCampaignAudit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/assets:
    get:
      tags: [Inventory]
      operationId: listAssets
      summary: List discovered assets
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: business_id
          in: query
          schema:
            type: string
            format: uuid
        - name: ad_account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: types
          in: query
          description: Comma-separated asset types.
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/AssetType"
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
        - name: search
          in: query
          description: Case-insensitive name or Meta-ID match.
          schema:
            type: string
      responses:
        "200":
          description: Paginated assets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/media:
    post:
      tags: [Media]
      operationId: uploadMedia
      summary: Store image or video media
      description: |
        Stores one file in the persistent local upload directory. Supplying a
        connection/account scopes ownership but does not upload the file to
        Meta. A publishing batch uploads bound media independently to each
        selected account.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file, kind]
              properties:
                file:
                  type: string
                  format: binary
                connection_id:
                  type: string
                  format: uuid
                ad_account_id:
                  type: string
                  format: uuid
                kind:
                  $ref: "#/components/schemas/MediaKind"
      responses:
        "201":
          description: File stored and ready for use in a batch media binding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Upload exceeds the configured request limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [Media]
      operationId: listMedia
      summary: List uploaded media
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: connection_id
          in: query
          schema:
            type: string
            format: uuid
        - name: ad_account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/MediaKind"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MediaStatus"
      responses:
        "200":
          description: Paginated media.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/media/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Media]
      operationId: getMedia
      summary: Get one media item
      responses:
        "200":
          description: Media item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/capabilities:
    get:
      tags: [Capabilities]
      operationId: getCapabilities
      summary: Get typed campaign-building capabilities
      description: |
        Returns the vocabulary modeled by this API. Meta can still reject a
        combination based on account capabilities, permissions, policy,
        country, billing, or API-version rules.
      responses:
        "200":
          description: Campaign-building capabilities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/batches/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Batches]
      operationId: getBatch
      summary: Get one publishing batch
      responses:
        "200":
          description: Batch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/batches/{id}/results:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Batches]
      operationId: listBatchResults
      summary: List per-account results for a batch
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: statuses
          in: query
          description: Comma-separated per-account statuses.
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/BatchAccountStatus"
      responses:
        "200":
          description: Paginated account results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchAccountResultPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/guards:
    get:
      tags: [Guards]
      summary: List checkpoint guards in the caller's tenant
      operationId: listGuards
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: batch_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Guard page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CampaignGuard" }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Exact value configured in `INTERNAL_API_TOKEN`.
...
  responses:
...
    NotFound:
      description: The requested local resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    OAuthCallbackResult:
      type: object
      additionalProperties: false
      required: [status, connection]
      properties:
        status:
          type: string
          enum: [connected]
        connection:
          $ref: "#/components/schemas/OAuthCompletion"
...
    ConnectionStatus:
      type: string
      enum: [active, expired, revoked, disconnected, error]
...